requirements.txt
streamlit
mcp-atlassian
transformers
torch
accelerate
python-dotenv